Move zed schema compile out of preview
